Fix gcc compiler warnings

+++ src/core/resource_manager.h
@@ -616,7 +616,7 @@ class ResourceManager {
           // this loop implements work stealing from other threads if there
           // are imbalances.
           for (int n = 0; n < p_numa_nodes; n++) {
-            uint64_t current_nid = (nid + n) % p_numa_nodes;
+            int current_nid = (nid + n) % p_numa_nodes;
             for (int thread_cnt = 0; thread_cnt < p_max_threads; thread_cnt++) {
               // uint64_t thread_cnt = 0;
               uint64_t current_tid = (tid + thread_cnt) % p_max_threads;
